+++ lambda-framework-repository/src/main/java/org/lambda/framework/repository/operation/ReactiveUnifyPagingRepositoryOperation.java
@@ -7,9 +7,9 @@ import static java.lang.Math.ceil;
 public interface ReactiveUnifyPagingRepositoryOperation {
 
 
-    default <Condition,Entity>Mono<Paged<Entity>> find(Integer page, Integer size,Condition condition,UnifyPagingOperation<Entity> operation){
-        return operation.count().switchIfEmpty(Mono.just(Integer.valueOf(0))).flatMap(e->{
-            return Mono.just(Paged.<Entity>builder().page(page).size(size).total(e).pages(((int)ceil(e/size))).build());
+    default <Condition,Entity>Mono<Paged<Entity>> find(Long page, Long size,Condition condition,UnifyPagingOperation<Entity> operation){
+        return operation.count().switchIfEmpty(Mono.just(Long.valueOf(0))).flatMap(e->{
+            return Mono.just(Paged.<Entity>builder().page(page).size(size).total(e).pages((long)(ceil(e/size))).build());
         }).flatMap(e->{
             return operation.<Entity>query().collectList().flatMap(records->{
                 e.setRecords(records);
