+++ lambda-framework-repository/src/main/java/org/lambda/framework/repository/operation/UnifyPagingOperation.java
@@ -4,7 +4,7 @@ import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
 public interface UnifyPagingOperation<Entity> {
-    public Mono<Integer> count();
+    public Mono<Long> count();
 
     public Flux<Entity> query();
 }
